docs for forecasting task (#443)
* docs for forecasting task

* avoid directly import extra dependencies

* Update docs/dev.rst


* make ForecastingDependenciesNotInstalledError a str message

* make ForecastingDependenciesNotInstalledError a str message

* update readme and examples

* add explanation for univariant models in example
