+++ docs/manual.rst
@@ -18,23 +18,34 @@ Examples
 ========
 * `Classification <examples/20_basics/example_tabular_classification.html>`_
 * `Regression <examples/20_basics/example_tabular_regression.html>`_
+* `Forecasting <examples/20_basic/example_time_series_forecasting.html>`_
 * `Customizing the search space <examples/40_advanced/example_custom_configuration_space.html>`_
 * `Changing the resampling strategy <examples/40_advanced/example_resampling_strategy.html>`_
 * `Visualizing the results <examples/40_advanced/example_visualization.html>`_
 
 Data validation
 ===============
-For tabular tasks, *Auto-PyTorch* uses a feature and target validator on the input feature set and target set respectively. 
+For **tabular tasks**, *Auto-PyTorch* uses a feature and target validator on the input feature set and target set respectively.
 
 The feature validator checks whether the data is supported by *Auto-PyTorch* or not. Additionally, a sklearn column transformer
 is also used which imputes and ordinally encodes the categorical columns of the dataset. This ensures
 that no unseen category is found while fitting the data.
 
-The target validator applies a label encoder on the target column. 
+The target validator applies a label encoder on the target column.
+
+For **time series forecasting tasks**, besides the functions described above, time series forecasting validators will also
+check the information specify for time series forecasting tasks: it checks
+    * The index of the series that each data point belongs to
+    * if the dataset is uni-variant (only targets information is contained in the datasets)
+    * the sample frequency of the datasets
+    * the static features in the dataset, i.e., features that contain only one value within each series
+
+Time Series forecasting validator then transforms the features and targets into a `pd.DataFrame <https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.html>`_
+whose index is applied to identify the series that the time step belongs to.
 
 Data Preprocessing
 ==================
-The tabular preprocessing pipeline in *Auto-PyTorch* consists of 
+The **tabular preprocessing pipeline** in *Auto-PyTorch* consists of
 
 #. `Imputation <https://github.com/automl/Auto-PyTorch/tree/development/autoPyTorch/pipeline/components/preprocessing/tabular_preprocessing/imputation>`_
 #. `Encoding <https://github.com/automl/Auto-PyTorch/tree/development/autoPyTorch/pipeline/components/preprocessing/tabular_preprocessing/encoding>`_
@@ -47,7 +58,24 @@ The tabular preprocessing pipeline in *Auto-PyTorch* consists of
 Along with the choices, their corresponding hyperparameters are also tuned. A sklearn ColumnTransformer is
 created which includes a categorical pipeline and a numerical pipeline. These pipelines are made up of the 
 relevant preprocessors chosen in the previous steps. The column transformer is compatible with `torchvision transforms <https://pytorch.org/vision/stable/transforms.html>`_
-and is therefore passed to the DataLoader. 
+and is therefore passed to the DataLoader.
+
+**time series forecasting pipeline** has two sorts of setup:
+
+- Uni-variant model only requires target transformations. They include *1:
+    #. `Target Imputation <https://github.com/automl/Auto-PyTorch/tree/development/autoPyTorch/pipeline/components/preprocessing/time_series_preprocessing/imputation/>`_
+        Choice of `linear`, `nearest`, `constant_zero`, `bfill` and `ffill`
+- Multi-variant model contains target transformations (see above) and feature transformation. They include
+    #. `Imputation <https://github.com/automl/Auto-PyTorch/tree/development/autoPyTorch/pipeline/components/preprocessing/time_series_preprocessing/imputation>`_
+         Choice of `linear`, `nearest`, `constant_zero`, `bfill` and `ffill`
+    #. `Scaling <https://github.com/automl/Auto-PyTorch/tree/development/autoPyTorch/pipeline/components/preprocessing/time_series_preprocessing/scaling>`_
+        Choice of `standard`, `min_max`, `max_abs`, `mean_abs`, or no transformation *2
+    #. `Encoding <https://github.com/automl/Auto-PyTorch/tree/development/autoPyTorch/pipeline/components/preprocessing/time_series_preprocessing/encoding>`_
+        Choice of `OneHotEncoder` or no encoding.
+
+*1 Target scaling is considered as part of `setup <https://github.com/automl/Auto-PyTorch/tree/development/autoPyTorch/pipeline/components/setup>`_ and the transform is done within each batch iteration
+
+*2 Scaling is transformed within each series
 
 Resource Allocation
 ===================
